+++ shared/messages/input/Image.h
@@ -41,12 +41,12 @@ namespace messages {
                 uint8_t cr;
             };
 
-            Image(size_t width, size_t height, NUClear::clock::time_point, std::vector<uint8_t>&& data);
+            Image(uint width, uint height, NUClear::clock::time_point, std::vector<uint8_t>&& data);
 
-            Pixel operator()(size_t x, size_t y) const;
+            Pixel operator()(uint x, uint y) const;
 
-            size_t width;
-            size_t height;
+            uint width;
+            uint height;
             NUClear::clock::time_point timestamp;
 
             // Returns the raw data that this is using
